requirements.txt
opencv-python
insightface==0.7.3
onnxruntime
numpy
faiss-cpu
tqdm
